requirements.txt
#############################
# Telegram Bot Framework
#############################
aiogram==3.4.1


#############################
# Async HTTP
#############################
aiohttp==3.9.3


#############################
# Database (PostgreSQL)
#############################
psycopg2-binary==2.9.9


#############################
# Redis / RQ
#############################
redis==5.0.1
rq==1.15.1
rq-scheduler==0.13.1


#############################
# OpenAI ChatGPT API
#############################
openai==1.12.0


#############################
# YooKassa (payments)
#############################
yookassa==3.1.0


#############################
# Environment variables
#############################
python-dotenv==1.0.1


#############################
# Utilities
#############################
pydantic==2.6.1
